WoT restructure D3: overlay/* -> *Overlay.ts + overlays.ts -> overlay/overlays.ts
