+++ handsontable/src/3rdparty/walkontable/src/overlay/bottomInlineStartCornerOverlay.ts
