+++ handsontable/src/3rdparty/walkontable/src/overlay/bottomOverlay.ts
