+++ handsontable/src/3rdparty/walkontable/src/overlay/index.ts
@@ -1,7 +1,7 @@
-export { BottomInlineStartCornerOverlay } from './bottomInlineStartCorner';
-export { BottomOverlay } from './bottom';
-export { InlineStartOverlay } from './inlineStart';
+export { BottomInlineStartCornerOverlay } from './bottomInlineStartCornerOverlay';
+export { BottomOverlay } from './bottomOverlay';
+export { InlineStartOverlay } from './inlineStartOverlay';
 export { Overlay } from './_base';
-export { TopInlineStartCornerOverlay } from './topInlineStartCorner';
-export { TopOverlay } from './top';
+export { TopInlineStartCornerOverlay } from './topInlineStartCornerOverlay';
+export { TopOverlay } from './topOverlay';
 export * from './constants';
