+++ handsontable/src/3rdparty/walkontable/src/overlay/inlineStartOverlay.ts
