+++ handsontable/src/3rdparty/walkontable/src/overlay/topInlineStartCornerOverlay.ts
