+++ handsontable/src/3rdparty/walkontable/src/overlay/topOverlay.ts
