+++ handsontable/src/3rdparty/walkontable/src/wire.ts
@@ -18,7 +18,7 @@ import type { WalkontableInstance } from './types';
 import type { default as Settings } from './settings';
 import type { default as Table } from './table/baseTable';
 import type { default as Viewport } from './viewport/viewport';
-import type { default as Overlays } from './overlays';
+import type { default as Overlays } from './overlay/overlays';
 import type { SelectionManager } from './selection/manager';
 import type { Overlay } from './overlay/_base';
 import type { GeometryReader } from './domMeasure/geometryReader';
